refactor: rename registration components and schemas for clarity, removing old registration forms and updating references to new personal registration form
