+++ producer/src/app/(protected)/perfil/comercial/components/CommercialInfoForm.tsx
@@ -3,7 +3,7 @@ import CustomInput from "@shared/components/CustomInput";
 import Image from "next/image";
 import { Control, UseFormRegister, UseFormGetValues, FieldErrors, Controller } from "react-hook-form";
 import { HiOutlinePencil } from "react-icons/hi";
-import { ChangeRegistrationSchema } from '@shared/schemas/change-registration';
+import { ChangeRegistrationSchema } from '@shared/schemas/change-personal-registration';
 
 interface CommercialInfoFormProps {
   photo: string;
