+++ producer/src/app/(protected)/perfil/components/ChangePersonalRegistrationForm/index.ts
@@ -1,0 +1,4 @@
+import ChangePersonalRegistrationForm from "./ChangePersonalRegistrationForm";
+import { useChangePersonalRegistrationForm } from "./useChangePersonalRegistrationForm";
+
+export { ChangePersonalRegistrationForm, useChangePersonalRegistrationForm };
