--- producer/src/app/(protected)/perfil/components/ChangeRegistrationForm/index.ts
@@ -1,4 +1,0 @@
-import ChangeRegistrationForm from "./ChangeRegistrationForm";
-import { useChangeRegistrationForm } from "./useChangeRegistrationForm";
-
-export { ChangeRegistrationForm, useChangeRegistrationForm };
