--- producer/src/app/alterar-cadastro/components/ChangePasswordForm/index.ts
@@ -1,4 +1,0 @@
-import ChangePasswordForm from "./ChangePasswordForm";
-import { useChangePasswordForm } from "./useChangePasswordForm";
-
-export { ChangePasswordForm, useChangePasswordForm };
