--- producer/src/app/alterar-cadastro/components/index.ts
@@ -1,6 +1,0 @@
-import {
-  ChangePasswordForm,
-  useChangePasswordForm,
-} from "./ChangePasswordForm";
-
-export { ChangePasswordForm, useChangePasswordForm };
